added new command line options for headless modes + several bug fixes

+++ Pladipus-search/src/main/java/com/compomics/pladipus/search/processsteps/SearchSetupStep.java
@@ -3,21 +3,15 @@ package com.compomics.pladipus.search.processsteps;
 import com.compomics.pladipus.core.control.util.PladipusFileDownloadingService;
 import com.compomics.pladipus.core.control.util.ZipUtils;
 import com.compomics.pladipus.core.model.processing.ProcessingStep;
-import com.compomics.util.experiment.identification.Advocate;
+import com.compomics.util.experiment.biology.taxonomy.SpeciesFactory;
 import com.compomics.util.experiment.identification.identification_parameters.SearchParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.AndromedaParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.CometParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.DirecTagParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.MsAmandaParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.MsgfParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.MyriMatchParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.OmssaParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.PNovoParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.PepnovoParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.TideParameters;
-import com.compomics.util.experiment.identification.identification_parameters.tool_specific.XtandemParameters;
+import com.compomics.util.preferences.IdentificationParameters;
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.net.URISyntaxException;
+import javax.xml.stream.XMLStreamException;
+import org.apache.commons.httpclient.URIException;
 import org.apache.commons.io.FileUtils;
 import org.apache.log4j.Logger;
 
@@ -43,7 +37,7 @@ public class SearchSetupStep extends ProcessingStep {
 
     @Override
     public boolean doAction() throws Exception {
-        if (tempResources.exists()) {
+        if (!parameters.containsKey("skip_cleaning") && tempResources.exists()) {
             for (File aFile : tempResources.listFiles()) {
                 if (aFile.exists()) {
                     if (aFile.isFile()) {
@@ -67,22 +61,23 @@ public class SearchSetupStep extends ProcessingStep {
         String fastaPath = parameters.get("fasta_file");
 
         //fix for older files that lack identification parameters
-        if (inputPath.toLowerCase().endsWith(".mgf") || inputPath.toLowerCase().endsWith(".mgf.zip")) {
-            File downloadFile = PladipusFileDownloadingService.downloadFile(inputPath, tempResources);
-            downloadFile.deleteOnExit();
-            String inputFile = downloadFile.getAbsolutePath();
+        if (!inputPath.equalsIgnoreCase(tempResources.getAbsolutePath())) {
+            if (inputPath.toLowerCase().endsWith(".mgf") || inputPath.toLowerCase().endsWith(".mgf.zip")) {
+                File downloadFile = PladipusFileDownloadingService.downloadFile(inputPath, tempResources);
+                downloadFile.deleteOnExit();
+                String inputFile = downloadFile.getAbsolutePath();
 
-            if (inputPath.toLowerCase().endsWith(".zip")) {
-                LOGGER.debug("Unzipping input");
-                ZipUtils.unzipArchive(new File(inputFile), tempResources);
-                LOGGER.debug("Done unzipping...");
+                if (inputPath.toLowerCase().endsWith(".zip")) {
+                    LOGGER.debug("Unzipping input");
+                    ZipUtils.unzipArchive(new File(inputFile), tempResources);
+                    LOGGER.debug("Done unzipping...");
+                }
+                parameters.put("spectrum_files", inputFile.replace(".zip", ""));
+            } else {
+                parameters.put("spectrum_files", PladipusFileDownloadingService.downloadFolder(inputPath, tempResources).getAbsolutePath());
             }
-            parameters.put("spectrum_files", inputFile.replace(".zip", ""));
-        } else {
-            parameters.put("spectrum_files", PladipusFileDownloadingService.downloadFolder(inputPath, tempResources).getAbsolutePath());
         }
         LOGGER.debug("Got input files " + parameters.get("spectrum_files"));
-
         //generate a repo folder for fasta files...
         //clear the repository save for the current fasta (temporary solution)
         //TODO refactor that it deletes complete runs only
@@ -98,24 +93,28 @@ public class SearchSetupStep extends ProcessingStep {
         }
         if (!fastaAlreadyExists) {
             fastaFile = PladipusFileDownloadingService.downloadFile(fastaPath, fasta_repo, fastaName);
-        }else{
-            for(File aFile:fasta_repo.listFiles()){
-                if(!aFile.getName().toLowerCase().contains(fastaName.toLowerCase())){
+        } else {
+            for (File aFile : fasta_repo.listFiles()) {
+                if (!aFile.getName().toLowerCase().contains(fastaName.toLowerCase())) {
                     aFile.delete();
                 }
             }
         }
-            
+
         if (fastaFile != null) {
             parameters.put("fasta_file", fastaFile.getAbsolutePath());
             LOGGER.debug("Got fasta file " + parameters.get("fasta_file"));
             //get and update parameters
-            File paramFile = PladipusFileDownloadingService.downloadFile(paramPath, tempResources);
-            paramFile.deleteOnExit();
-            SearchParameters identificationParameters = SearchParameters.getIdentificationParameters(paramFile);
-            SearchParameters updatedIdentificationParameters = updateAlgorithmSettings(identificationParameters, fastaFile);
-            SearchParameters.saveIdentificationParameters(updatedIdentificationParameters, paramFile);
-            parameters.put("id_params", paramFile.getAbsolutePath());
+            File paramFile;
+            if (!paramPath.contains(tempResources.getAbsolutePath())) {
+                paramFile = PladipusFileDownloadingService.downloadFile(paramPath, tempResources);
+                parameters.put("id_params", paramFile.getAbsolutePath());
+            }
+            paramFile = new File(parameters.get("id_params"));
+            SearchParameters sparameters = SearchParameters.getIdentificationParameters(paramFile);
+            IdentificationParameters updatedIdentificationParameters = updateAlgorithmSettings(sparameters, fastaFile);
+            IdentificationParameters.saveIdentificationParameters(updatedIdentificationParameters, paramFile);
+
             LOGGER.info("Got identification parameters " + parameters.get("id_params"));
 
             //output
@@ -133,77 +132,12 @@ public class SearchSetupStep extends ProcessingStep {
         return "Initialisation of the search process";
     }
 
-    public SearchParameters updateAlgorithmSettings(SearchParameters searchParameters, File fasta) {
-        SearchParameters tempSearchParameters = new SearchParameters();
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.omssa.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.omssa.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.omssa.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.omssa.getIndex(), new OmssaParameters());
-        }
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.xtandem.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.xtandem.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.xtandem.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.xtandem.getIndex(), new XtandemParameters());
-        }
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.msgf.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.msgf.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.msgf.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.msgf.getIndex(), new MsgfParameters());
-        }
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.msAmanda.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.msAmanda.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.msAmanda.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.msAmanda.getIndex(), new MsAmandaParameters());
-        }
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.myriMatch.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.myriMatch.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.myriMatch.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.myriMatch.getIndex(), new MyriMatchParameters());
-        }
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.comet.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.comet.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.comet.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.comet.getIndex(), new CometParameters());
-        }
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.tide.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.tide.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.tide.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.tide.getIndex(), new TideParameters());
-        }
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.andromeda.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.andromeda.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.andromeda.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.andromeda.getIndex(), new AndromedaParameters());
-        }
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.pepnovo.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.pepnovo.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.pepnovo.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.pepnovo.getIndex(), new PepnovoParameters());
-        }
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.direcTag.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.direcTag.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.direcTag.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.direcTag.getIndex(), new DirecTagParameters());
-        }
-        if (searchParameters.getIdentificationAlgorithmParameter(Advocate.pNovo.getIndex()) != null) {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.pNovo.getIndex(), searchParameters.getIdentificationAlgorithmParameter(Advocate.pNovo.getIndex()));
-        } else {
-            tempSearchParameters.setIdentificationAlgorithmParameter(Advocate.pNovo.getIndex(), new PNovoParameters());
-        }
-        //set the fasta file
-        tempSearchParameters.setFastaFile(fasta);
-        //
-        tempSearchParameters.setEnzyme(searchParameters.getEnzyme());
-        tempSearchParameters.setPrecursorAccuracy(searchParameters.getPrecursorAccuracy());
-        tempSearchParameters.setFragmentIonAccuracy(searchParameters.getFragmentIonAccuracy());
-        tempSearchParameters.setPrecursorAccuracyType(searchParameters.getPrecursorAccuracyType());
-        tempSearchParameters.setFragmentAccuracyType(searchParameters.getFragmentAccuracyType());
-        tempSearchParameters.setPtmSettings(searchParameters.getPtmSettings());
-        tempSearchParameters.setnMissedCleavages(searchParameters.getnMissedCleavages());
-        tempSearchParameters.setMaxChargeSearched(searchParameters.getMaxChargeSearched());
-        tempSearchParameters.setMinChargeSearched(searchParameters.getMinChargeSearched());
-     //   System.out.println(tempSearchParameters.toString());
-
-        return tempSearchParameters;
+    public IdentificationParameters updateAlgorithmSettings(SearchParameters searchParameters, File fasta) throws URIException, IOException, XMLStreamException, URISyntaxException {
+        System.out.println("Updating the algorithm settings and setting the fasta file...");
+        searchParameters.setFastaFile(fasta);
+        SpeciesFactory.getInstance().initiate(new SearchGUIStep().getJar().getParentFile().getAbsolutePath());
+        IdentificationParameters temp = new IdentificationParameters(searchParameters);
+        return temp;
     }
+
 }
